Disable trigram indexing in SQLite search

+++ Veriado.Infrastructure/Search/SqliteFts5Transactional.cs
@@ -28,6 +28,8 @@ internal sealed class SqliteFts5Transactional
         _trigramBuilder = trigramBuilder ?? throw new ArgumentNullException(nameof(trigramBuilder));
     }
 
+    private bool IsTrigramIndexEnabled => _trigramOptions.Fields is { Length: > 0 } fields && fields.Length > 0;
+
     public async Task<long?> IndexAsync(
         SearchDocument document,
         SqliteConnection connection,
@@ -59,7 +61,12 @@ internal sealed class SqliteFts5Transactional
                     cancellationToken)
                 .ConfigureAwait(false);
             var searchRowId = await EnsureMapRowIdAsync("file_search_map", document.FileId, connection, transaction, cancellationToken).ConfigureAwait(false);
-            var trigramRowId = await EnsureMapRowIdAsync("file_trgm_map", document.FileId, connection, transaction, cancellationToken).ConfigureAwait(false);
+            long? trigramRowId = null;
+            if (IsTrigramIndexEnabled)
+            {
+                trigramRowId = await EnsureMapRowIdAsync("file_trgm_map", document.FileId, connection, transaction, cancellationToken)
+                    .ConfigureAwait(false);
+            }
 
             await using (var delete = connection.CreateCommand())
             {
@@ -82,23 +89,26 @@ internal sealed class SqliteFts5Transactional
                 await insert.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
             }
 
-            var trigramText = BuildTrigramText(document);
-
-            await using (var deleteTrgm = connection.CreateCommand())
+            if (trigramRowId.HasValue)
             {
-                deleteTrgm.Transaction = transaction;
-                deleteTrgm.CommandText = "INSERT INTO file_trgm(file_trgm, rowid) VALUES ('delete', $rowid);";
-                deleteTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId;
-                await deleteTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
-            }
+                var trigramText = BuildTrigramText(document);
 
-            await using (var insertTrgm = connection.CreateCommand())
-            {
-                insertTrgm.Transaction = transaction;
-                insertTrgm.CommandText = "INSERT INTO file_trgm(rowid, trgm) VALUES ($rowid, $trgm);";
-                insertTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId;
-                insertTrgm.Parameters.AddWithValue("$trgm", trigramText);
-                await insertTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
+                await using (var deleteTrgm = connection.CreateCommand())
+                {
+                    deleteTrgm.Transaction = transaction;
+                    deleteTrgm.CommandText = "INSERT INTO file_trgm(file_trgm, rowid) VALUES ('delete', $rowid);";
+                    deleteTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId.Value;
+                    await deleteTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
+                }
+
+                await using (var insertTrgm = connection.CreateCommand())
+                {
+                    insertTrgm.Transaction = transaction;
+                    insertTrgm.CommandText = "INSERT INTO file_trgm(rowid, trgm) VALUES ($rowid, $trgm);";
+                    insertTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId.Value;
+                    insertTrgm.Parameters.AddWithValue("$trgm", trigramText);
+                    await insertTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
+                }
             }
 
             if (enlistJournal && journalId.HasValue)
@@ -159,14 +169,18 @@ internal sealed class SqliteFts5Transactional
                 await deleteFts.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
             }
 
-            var trigramRowId = await TryGetRowIdAsync("file_trgm_map", fileId, connection, transaction, cancellationToken).ConfigureAwait(false);
-            if (trigramRowId.HasValue)
+            if (IsTrigramIndexEnabled)
             {
-                await using var deleteTrgm = connection.CreateCommand();
-                deleteTrgm.Transaction = transaction;
-                deleteTrgm.CommandText = "INSERT INTO file_trgm(file_trgm, rowid) VALUES ('delete', $rowid);";
-                deleteTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId.Value;
-                await deleteTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
+                var trigramRowId = await TryGetRowIdAsync("file_trgm_map", fileId, connection, transaction, cancellationToken)
+                    .ConfigureAwait(false);
+                if (trigramRowId.HasValue)
+                {
+                    await using var deleteTrgm = connection.CreateCommand();
+                    deleteTrgm.Transaction = transaction;
+                    deleteTrgm.CommandText = "INSERT INTO file_trgm(file_trgm, rowid) VALUES ('delete', $rowid);";
+                    deleteTrgm.Parameters.Add("$rowid", SqliteType.Integer).Value = trigramRowId.Value;
+                    await deleteTrgm.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
+                }
             }
 
             await using (var deleteSearchMap = connection.CreateCommand())
@@ -177,8 +191,9 @@ internal sealed class SqliteFts5Transactional
                 await deleteSearchMap.ExecuteNonQueryAsync(cancellationToken).ConfigureAwait(false);
             }
 
-            await using (var deleteTrgmMap = connection.CreateCommand())
+            if (IsTrigramIndexEnabled)
             {
+                await using var deleteTrgmMap = connection.CreateCommand();
                 deleteTrgmMap.Transaction = transaction;
                 deleteTrgmMap.CommandText = "DELETE FROM file_trgm_map WHERE file_id = $fileId;";
                 deleteTrgmMap.Parameters.Add("$fileId", SqliteType.Blob).Value = fileKey;
